Add batch to dataset if user manually sets feeling.
If the batch size is at least some minimum size (currently 1000) and the
user sets checks in, then this batch will be added to the dataset. While
this batch may seem small, feelings manually set likely more precisely
represent the feeling of the set. This also allows the user to set a
feeling before the timeout occurs.